requirements.txt
aioamqp-beta==0.0.1
aiomysql==0.0.20
aioredis==1.2.0
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
importlib-metadata==0.18
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
motor==2.0.0
packaging==19.0
pamqp==2.3.0
peewee==2.10.2
peewee-async==0.5.12
pluggy==0.12.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pymongo==3.8.0
PyMySQL==0.9.2
pyparsing==2.4.1.1
pytest==5.0.1
pytest-asyncio==0.10.0
PyYAML==5.1.1
redis==3.2.1
six==1.12.0
tornado==6.0.3
tornado-swagger==1.2.0
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.2
